requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
cssselect==1.1.0
dateparser==1.1.1
demjson3==3.0.5
facebook-scraper==0.2.57
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.0
numpy==1.23.0
pandas==1.4.3
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vaderSentiment==3.3.2
w3lib==1.22.0
websockets==10.3
zipp==3.8.0
